requirements.txt
asgiref==3.7.2
Django==4.2.7
pip==23.3.1
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
python-decouple==3.8
django-cors-headers==4.3.1
djangorestframework==3.14.0
psycopg2-binary==2.9.5
gunicorn==20.1.0
django-storages==1.13.2
azure-core==1.26.3
azure-storage-blob==12.14.1